version changed

+++ eth/gasprice/feehistory_test.go
@@ -28,8 +28,8 @@ import (
 func TestFeeHistory(t *testing.T) {
 	var cases = []struct {
 		pending             bool
-		maxHeader, maxBlock uint64
-		count               uint64
+		maxHeader, maxBlock int
+		count               int
 		last                rpc.BlockNumber
 		percent             []float64
 		expFirst            uint64
